+++ Source/CesiumEditor/Private/CesiumIonSession.cpp
@@ -1,114 +1,223 @@
 // Copyright 2020-2021 CesiumGS, Inc. and Contributors
 
 #include "CesiumIonSession.h"
+#include "CesiumEditor.h"
 #include "CesiumEditorSettings.h"
+#include "CesiumIonServer.h"
 #include "CesiumRuntimeSettings.h"
 #include "CesiumSourceControl.h"
+#include "CesiumUtility/Uri.h"
+#include "FileHelpers.h"
 #include "HAL/PlatformProcess.h"
 #include "Misc/App.h"
 
 using namespace CesiumAsync;
 using namespace CesiumIonClient;
 
+namespace {
+
+template <typename T> void logResponseErrors(const Response<T>& response) {
+  if (!response.errorCode.empty() && !response.errorMessage.empty()) {
+    UE_LOG(
+        LogCesiumEditor,
+        Error,
+        TEXT("%s (Code %s)"),
+        UTF8_TO_TCHAR(response.errorMessage.c_str()),
+        UTF8_TO_TCHAR(response.errorCode.c_str()));
+  } else if (!response.errorCode.empty()) {
+    UE_LOG(
+        LogCesiumEditor,
+        Error,
+        TEXT("Code %s"),
+        UTF8_TO_TCHAR(response.errorCode.c_str()));
+  } else if (!response.errorMessage.empty()) {
+    UE_LOG(
+        LogCesiumEditor,
+        Error,
+        TEXT("%s"),
+        UTF8_TO_TCHAR(response.errorMessage.c_str()));
+  }
+}
+
+void logResponseErrors(const std::exception& exception) {
+  UE_LOG(
+      LogCesiumEditor,
+      Error,
+      TEXT("Exception: %s"),
+      UTF8_TO_TCHAR(exception.what()));
+}
+
+} // namespace
+
 CesiumIonSession::CesiumIonSession(
     CesiumAsync::AsyncSystem& asyncSystem,
-    const std::shared_ptr<CesiumAsync::IAssetAccessor>& pAssetAccessor)
+    const std::shared_ptr<CesiumAsync::IAssetAccessor>& pAssetAccessor,
+    TWeakObjectPtr<UCesiumIonServer> pServer)
     : _asyncSystem(asyncSystem),
       _pAssetAccessor(pAssetAccessor),
+      _pServer(pServer),
       _connection(std::nullopt),
       _profile(std::nullopt),
       _assets(std::nullopt),
       _tokens(std::nullopt),
+      _defaults(std::nullopt),
       _isConnecting(false),
       _isResuming(false),
       _isLoadingProfile(false),
       _isLoadingAssets(false),
       _isLoadingTokens(false),
+      _isLoadingDefaults(false),
       _loadProfileQueued(false),
       _loadAssetsQueued(false),
       _loadTokensQueued(false),
+      _loadDefaultsQueued(false),
       _authorizeUrl() {}
 
 void CesiumIonSession::connect() {
-  if (this->isConnecting() || this->isConnected() || this->isResuming()) {
+  if (!this->_pServer.IsValid() || this->isConnecting() ||
+      this->isConnected() || this->isResuming()) {
     return;
   }
 
+  UCesiumIonServer* pServer = this->_pServer.Get();
+
   this->_isConnecting = true;
 
-  Connection::authorize(
-      this->_asyncSystem,
-      this->_pAssetAccessor,
-      "Cesium for Unreal",
-      190,
-      "/cesium-for-unreal/oauth2/callback",
-      {"assets:list",
-       "assets:read",
-       "profile:read",
-       "tokens:read",
-       "tokens:write",
-       "geocode"},
-      [this](const std::string& url) {
-        this->_authorizeUrl = url;
-        FPlatformProcess::LaunchURL(
-            UTF8_TO_TCHAR(this->_authorizeUrl.c_str()),
-            NULL,
-            NULL);
-      })
-      .thenInMainThread([this](CesiumIonClient::Connection&& connection) {
-        this->_isConnecting = false;
-        this->_connection = std::move(connection);
+  std::string ionServerUrl = TCHAR_TO_UTF8(*pServer->ServerUrl);
 
-        UCesiumEditorSettings* pSettings =
-            GetMutableDefault<UCesiumEditorSettings>();
-        pSettings->UserAccessToken =
-            UTF8_TO_TCHAR(this->_connection.value().getAccessToken().c_str());
+  Future<std::optional<std::string>> futureApiUrl =
+      !pServer->ApiUrl.IsEmpty()
+          ? this->_asyncSystem.createResolvedFuture<std::optional<std::string>>(
+                TCHAR_TO_UTF8(*pServer->ApiUrl))
+          : Connection::getApiUrl(
+                this->_asyncSystem,
+                this->_pAssetAccessor,
+                ionServerUrl);
 
-        CesiumSourceControl::PromptToCheckoutConfigFile(
-            pSettings->GetClass()->GetConfigName());
+  std::shared_ptr<CesiumIonSession> thiz = this->shared_from_this();
 
-        pSettings->SaveConfig();
+  std::move(futureApiUrl)
+      .thenInMainThread([ionServerUrl, thiz, pServer = this->_pServer](
+                            std::optional<std::string>&& ionApiUrl) {
+        if (!pServer.IsValid()) {
+          thiz->_isConnecting = false;
+          thiz->_connection = std::nullopt;
+          thiz->ConnectionUpdated.Broadcast();
+          return;
+        }
 
-        this->ConnectionUpdated.Broadcast();
-      })
-      .catchInMainThread([this](std::exception&& e) {
-        this->_isConnecting = false;
-        this->_connection = std::nullopt;
-        this->ConnectionUpdated.Broadcast();
+        if (!ionApiUrl) {
+          thiz->_isConnecting = false;
+          thiz->_connection = std::nullopt;
+          thiz->ConnectionUpdated.Broadcast();
+          UE_LOG(
+              LogCesiumEditor,
+              Error,
+              TEXT(
+                  "Failed to retrieve API URL from the config.json file at the specified Ion server URL: %s"),
+              UTF8_TO_TCHAR(ionServerUrl.c_str()));
+          return;
+        }
+
+        if (pServer->ApiUrl.IsEmpty()) {
+          pServer->ApiUrl = UTF8_TO_TCHAR(ionApiUrl->c_str());
+          pServer->Modify();
+          UEditorLoadingAndSavingUtils::SavePackages(
+              {pServer->GetPackage()},
+              true);
+        }
+
+        int64_t clientID = pServer->OAuth2ApplicationID;
+
+        Connection::authorize(
+            thiz->_asyncSystem,
+            thiz->_pAssetAccessor,
+            "Cesium for Unreal",
+            clientID,
+            "/cesium-for-unreal/oauth2/callback",
+            {"assets:list",
+             "assets:read",
+             "profile:read",
+             "tokens:read",
+             "tokens:write",
+             "geocode"},
+            [thiz](const std::string& url) {
+              thiz->_authorizeUrl = url;
+
+              thiz->_redirectUrl =
+                  CesiumUtility::Uri::getQueryValue(url, "redirect_uri");
+
+              FPlatformProcess::LaunchURL(
+                  UTF8_TO_TCHAR(thiz->_authorizeUrl.c_str()),
+                  NULL,
+                  NULL);
+            },
+            *ionApiUrl,
+            CesiumUtility::Uri::resolve(ionServerUrl, "oauth"))
+            .thenInMainThread([thiz](CesiumIonClient::Connection&& connection) {
+              thiz->_isConnecting = false;
+              thiz->_connection = std::move(connection);
+
+              UCesiumEditorSettings* pSettings =
+                  GetMutableDefault<UCesiumEditorSettings>();
+              pSettings->UserAccessTokenMap.Add(
+                  thiz->_pServer.Get(),
+                  UTF8_TO_TCHAR(
+                      thiz->_connection.value().getAccessToken().c_str()));
+              pSettings->Save();
+
+              thiz->ConnectionUpdated.Broadcast();
+
+              thiz->startQueuedLoads();
+            })
+            .catchInMainThread([thiz](std::exception&& e) {
+              thiz->_isConnecting = false;
+              thiz->_connection = std::nullopt;
+              thiz->ConnectionUpdated.Broadcast();
+            });
       });
 }
 
 void CesiumIonSession::resume() {
-  if (this->isConnecting() || this->isConnected() || this->isResuming()) {
+  if (!this->_pServer.IsValid() || this->isConnecting() ||
+      this->isConnected() || this->isResuming()) {
     return;
   }
 
   const UCesiumEditorSettings* pSettings = GetDefault<UCesiumEditorSettings>();
-  if (pSettings->UserAccessToken.IsEmpty()) {
+  const FString* pUserAccessToken =
+      pSettings->UserAccessTokenMap.Find(this->_pServer.Get());
+
+  if (!pUserAccessToken || pUserAccessToken->IsEmpty()) {
     // No existing session to resume.
     return;
   }
 
   this->_isResuming = true;
 
-  this->_connection = Connection(
+  std::shared_ptr<Connection> pConnection = std::make_shared<Connection>(
       this->_asyncSystem,
       this->_pAssetAccessor,
-      TCHAR_TO_UTF8(*GetDefault<UCesiumEditorSettings>()->UserAccessToken));
+      TCHAR_TO_UTF8(**pUserAccessToken),
+      TCHAR_TO_UTF8(*this->_pServer->ApiUrl));
+
+  std::shared_ptr<CesiumIonSession> thiz = this->shared_from_this();
 
   // Verify that the connection actually works.
-  this->_connection.value()
-      .me()
-      .thenInMainThread([this](Response<Profile>&& response) {
-        if (!response.value.has_value()) {
-          this->_connection.reset();
+  pConnection->me()
+      .thenInMainThread([thiz, pConnection](Response<Profile>&& response) {
+        logResponseErrors(response);
+        if (response.value.has_value()) {
+          thiz->_connection = std::move(*pConnection);
         }
-        this->_isResuming = false;
-        this->ConnectionUpdated.Broadcast();
+        thiz->_isResuming = false;
+        thiz->ConnectionUpdated.Broadcast();
+
+        thiz->startQueuedLoads();
       })
-      .catchInMainThread([this](std::exception&& e) {
-        this->_isResuming = false;
-        this->_connection.reset();
+      .catchInMainThread([thiz](std::exception&& e) {
+        logResponseErrors(e);
+        thiz->_isResuming = false;
       });
 }
 
@@ -117,18 +226,17 @@ void CesiumIonSession::disconnect() {
   this->_profile.reset();
   this->_assets.reset();
   this->_tokens.reset();
+  this->_defaults.reset();
 
   UCesiumEditorSettings* pSettings = GetMutableDefault<UCesiumEditorSettings>();
-  pSettings->UserAccessToken.Empty();
-
-  CesiumSourceControl::PromptToCheckoutConfigFile(
-      pSettings->GetClass()->GetConfigName());
-  pSettings->SaveConfig();
+  pSettings->UserAccessTokenMap.Remove(this->_pServer.Get());
+  pSettings->Save();
 
   this->ConnectionUpdated.Broadcast();
   this->ProfileUpdated.Broadcast();
   this->AssetsUpdated.Broadcast();
   this->TokensUpdated.Broadcast();
+  this->DefaultsUpdated.Broadcast();
 }
 
 void CesiumIonSession::refreshProfile() {
@@ -140,66 +248,116 @@ void CesiumIonSession::refreshProfile() {
   this->_isLoadingProfile = true;
   this->_loadProfileQueued = false;
 
+  std::shared_ptr<CesiumIonSession> thiz = this->shared_from_this();
+
   this->_connection->me()
-      .thenInMainThread([this](Response<Profile>&& profile) {
-        this->_isLoadingProfile = false;
-        this->_profile = std::move(profile.value);
-        this->ProfileUpdated.Broadcast();
-        this->refreshProfileIfNeeded();
+      .thenInMainThread([thiz](Response<Profile>&& profile) {
+        logResponseErrors(profile);
+        thiz->_isLoadingProfile = false;
+        thiz->_profile = std::move(profile.value);
+        thiz->ProfileUpdated.Broadcast();
+        if (thiz->_loadProfileQueued)
+          thiz->refreshProfile();
       })
-      .catchInMainThread([this](std::exception&& e) {
-        this->_isLoadingProfile = false;
-        this->_profile = std::nullopt;
-        this->ProfileUpdated.Broadcast();
-        this->refreshProfileIfNeeded();
+      .catchInMainThread([thiz](std::exception&& e) {
+        logResponseErrors(e);
+        thiz->_isLoadingProfile = false;
+        thiz->_profile = std::nullopt;
+        thiz->ProfileUpdated.Broadcast();
+        if (thiz->_loadProfileQueued)
+          thiz->refreshProfile();
       });
 }
 
 void CesiumIonSession::refreshAssets() {
   if (!this->_connection || this->_isLoadingAssets) {
+    this->_loadAssetsQueued = true;
     return;
   }
 
   this->_isLoadingAssets = true;
   this->_loadAssetsQueued = false;
 
+  std::shared_ptr<CesiumIonSession> thiz = this->shared_from_this();
+
   this->_connection->assets()
-      .thenInMainThread([this](Response<Assets>&& assets) {
-        this->_isLoadingAssets = false;
-        this->_assets = std::move(assets.value);
-        this->AssetsUpdated.Broadcast();
-        this->refreshAssetsIfNeeded();
+      .thenInMainThread([thiz](Response<Assets>&& assets) {
+        logResponseErrors(assets);
+        thiz->_isLoadingAssets = false;
+        thiz->_assets = std::move(assets.value);
+        thiz->AssetsUpdated.Broadcast();
+        if (thiz->_loadAssetsQueued)
+          thiz->refreshAssets();
       })
-      .catchInMainThread([this](std::exception&& e) {
-        this->_isLoadingAssets = false;
-        this->_assets = std::nullopt;
-        this->AssetsUpdated.Broadcast();
-        this->refreshAssetsIfNeeded();
+      .catchInMainThread([thiz](std::exception&& e) {
+        logResponseErrors(e);
+        thiz->_isLoadingAssets = false;
+        thiz->_assets = std::nullopt;
+        thiz->AssetsUpdated.Broadcast();
+        if (thiz->_loadAssetsQueued)
+          thiz->refreshAssets();
       });
 }
 
 void CesiumIonSession::refreshTokens() {
   if (!this->_connection || this->_isLoadingTokens) {
+    this->_loadTokensQueued = true;
     return;
   }
 
   this->_isLoadingTokens = true;
   this->_loadTokensQueued = false;
 
+  std::shared_ptr<CesiumIonSession> thiz = this->shared_from_this();
+
   this->_connection->tokens()
-      .thenInMainThread([this](Response<TokenList>&& tokens) {
-        this->_isLoadingTokens = false;
-        this->_tokens = tokens.value
+      .thenInMainThread([thiz](Response<TokenList>&& tokens) {
+        logResponseErrors(tokens);
+        thiz->_isLoadingTokens = false;
+        thiz->_tokens = tokens.value
                             ? std::make_optional(std::move(tokens.value->items))
                             : std::nullopt;
-        this->TokensUpdated.Broadcast();
-        this->refreshTokensIfNeeded();
+        thiz->TokensUpdated.Broadcast();
+        if (thiz->_loadTokensQueued)
+          thiz->refreshTokens();
       })
-      .catchInMainThread([this](std::exception&& e) {
-        this->_isLoadingTokens = false;
-        this->_tokens = std::nullopt;
-        this->TokensUpdated.Broadcast();
-        this->refreshTokensIfNeeded();
+      .catchInMainThread([thiz](std::exception&& e) {
+        logResponseErrors(e);
+        thiz->_isLoadingTokens = false;
+        thiz->_tokens = std::nullopt;
+        thiz->TokensUpdated.Broadcast();
+        if (thiz->_loadTokensQueued)
+          thiz->refreshTokens();
+      });
+}
+
+void CesiumIonSession::refreshDefaults() {
+  if (!this->_connection || this->_isLoadingDefaults) {
+    this->_loadDefaultsQueued = true;
+    return;
+  }
+
+  this->_isLoadingDefaults = true;
+  this->_loadDefaultsQueued = false;
+
+  std::shared_ptr<CesiumIonSession> thiz = this->shared_from_this();
+
+  this->_connection->defaults()
+      .thenInMainThread([thiz](Response<Defaults>&& defaults) {
+        logResponseErrors(defaults);
+        thiz->_isLoadingDefaults = false;
+        thiz->_defaults = std::move(defaults.value);
+        thiz->DefaultsUpdated.Broadcast();
+        if (thiz->_loadDefaultsQueued)
+          thiz->refreshDefaults();
+      })
+      .catchInMainThread([thiz](std::exception&& e) {
+        logResponseErrors(e);
+        thiz->_isLoadingDefaults = false;
+        thiz->_defaults = std::nullopt;
+        thiz->DefaultsUpdated.Broadcast();
+        if (thiz->_loadDefaultsQueued)
+          thiz->refreshDefaults();
       });
 }
 
@@ -238,6 +396,16 @@ const std::vector<CesiumIonClient::Token>& CesiumIonSession::getTokens() {
   }
 }
 
+const CesiumIonClient::Defaults& CesiumIonSession::getDefaults() {
+  static const CesiumIonClient::Defaults empty;
+  if (this->_defaults) {
+    return *this->_defaults;
+  } else {
+    this->refreshDefaults();
+    return empty;
+  }
+}
+
 bool CesiumIonSession::refreshProfileIfNeeded() {
   if (this->_loadProfileQueued || !this->_profile.has_value()) {
     this->refreshProfile();
@@ -257,6 +425,13 @@ bool CesiumIonSession::refreshTokensIfNeeded() {
     this->refreshTokens();
   }
   return this->isTokenListLoaded();
+}
+
+bool CesiumIonSession::refreshDefaultsIfNeeded() {
+  if (this->_loadDefaultsQueued || !this->_defaults.has_value()) {
+    this->refreshDefaults();
+  }
+  return this->isDefaultsLoaded();
 }
 
 Future<Response<Token>>
@@ -280,39 +455,42 @@ CesiumIonSession::findToken(const FString& token) const {
 
 namespace {
 
-Token tokenFromSettings() {
+Token tokenFromServer(UCesiumIonServer* pServer) {
   Token result;
-  result.token = TCHAR_TO_UTF8(
-      *GetDefault<UCesiumRuntimeSettings>()->DefaultIonAccessToken);
+
+  if (pServer) {
+    result.token = TCHAR_TO_UTF8(*pServer->DefaultIonAccessToken);
+  }
+
   return result;
 }
 
 Future<Token> getTokenFuture(const CesiumIonSession& session) {
-  if (!GetDefault<UCesiumRuntimeSettings>()
-           ->DefaultIonAccessTokenId.IsEmpty()) {
+  std::shared_ptr<const CesiumIonSession> pSession = session.shared_from_this();
+  TWeakObjectPtr<UCesiumIonServer> pServer = session.getServer();
+
+  if (pServer.IsValid() && !pServer->DefaultIonAccessTokenId.IsEmpty()) {
     return session.getConnection()
-        ->token(TCHAR_TO_UTF8(
-            *GetDefault<UCesiumRuntimeSettings>()->DefaultIonAccessTokenId))
-        .thenImmediately([](Response<Token>&& tokenResponse) {
+        ->token(TCHAR_TO_UTF8(*pServer->DefaultIonAccessTokenId))
+        .thenImmediately([pServer](Response<Token>&& tokenResponse) {
           if (tokenResponse.value) {
             return *tokenResponse.value;
           } else {
-            return tokenFromSettings();
+            return tokenFromServer(pServer.Get());
           }
         });
-  } else if (!GetDefault<UCesiumRuntimeSettings>()
-                  ->DefaultIonAccessToken.IsEmpty()) {
-    return session
-        .findToken(GetDefault<UCesiumRuntimeSettings>()->DefaultIonAccessToken)
-        .thenImmediately([](Response<Token>&& response) {
+  } else if (!pServer->DefaultIonAccessToken.IsEmpty()) {
+    return session.findToken(pServer->DefaultIonAccessToken)
+        .thenImmediately([pServer](Response<Token>&& response) {
           if (response.value) {
             return *response.value;
           } else {
-            return tokenFromSettings();
+            return tokenFromServer(pServer.Get());
           }
         });
   } else {
-    return session.getAsyncSystem().createResolvedFuture(tokenFromSettings());
+    return session.getAsyncSystem().createResolvedFuture(
+        tokenFromServer(pServer.Get()));
   }
 }
 
@@ -325,8 +503,7 @@ SharedFuture<Token> CesiumIonSession::getProjectDefaultTokenDetails() {
     // new token.
     if (this->_projectDefaultTokenDetailsFuture->isReady() &&
         this->_projectDefaultTokenDetailsFuture->wait().token !=
-            TCHAR_TO_UTF8(
-                *GetDefault<UCesiumRuntimeSettings>()->DefaultIonAccessToken)) {
+            TCHAR_TO_UTF8(*this->_pServer->DefaultIonAccessToken)) {
       this->_projectDefaultTokenDetailsFuture.reset();
     } else {
       return *this->_projectDefaultTokenDetailsFuture;
@@ -335,7 +512,7 @@ SharedFuture<Token> CesiumIonSession::getProjectDefaultTokenDetails() {
 
   if (!this->isConnected()) {
     return this->getAsyncSystem()
-        .createResolvedFuture(tokenFromSettings())
+        .createResolvedFuture(tokenFromServer(this->_pServer.Get()))
         .share();
   }
 
@@ -345,4 +522,15 @@ SharedFuture<Token> CesiumIonSession::getProjectDefaultTokenDetails() {
 
 void CesiumIonSession::invalidateProjectDefaultTokenDetails() {
   this->_projectDefaultTokenDetailsFuture.reset();
+}
+
+void CesiumIonSession::startQueuedLoads() {
+  if (this->_loadProfileQueued)
+    this->refreshProfile();
+  if (this->_loadAssetsQueued)
+    this->refreshAssets();
+  if (this->_loadTokensQueued)
+    this->refreshTokens();
+  if (this->_loadDefaultsQueued)
+    this->refreshDefaults();
 }