+++ Source/CesiumRuntime/Public/CesiumIonRasterOverlay.h
@@ -6,6 +6,8 @@
 #include "CoreMinimal.h"
 #include "CesiumIonRasterOverlay.generated.h"
 
+class UCesiumIonServer;
+
 /**
  * A raster overlay that uses an IMAGERY asset from Cesium ion.
  */
@@ -37,11 +39,20 @@ public:
    * endpoint can be specified.
    */
   UPROPERTY(
+      meta =
+          (DeprecatedProperty,
+           DeprecationMessage = "Use CesiumIonServer instead."))
+  FString IonAssetEndpointUrl_DEPRECATED;
+
+  /**
+   * The Cesium ion Server from which this raster overlay is loaded.
+   */
+  UPROPERTY(
       EditAnywhere,
       BlueprintReadWrite,
       Category = "Cesium",
       AdvancedDisplay)
-  FString IonAssetEndpointUrl;
+  UCesiumIonServer* CesiumIonServer;
 
   /**
    * Check if the Cesium ion token used to access this raster overlay is working